+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * fixed bug for sorting records in the admin panel;
 * fixed bugs and improved code for fields that must be unique;
 * improved some validation rules;
+* added tests for `TreeBehavior` class;
 * added tests for all validation classes.
 
 ### 2.14.12